README: Fix up the EBNF for the language a bit
Reformatting the EBNF code in the README, but also fix a few tiny
mistakes. Like previously the following was allowed

```
IF test < 0 THEN
  RETURN
FI
```

A bit of a problem now is that the definitions for statements inside a
function are different from ones outside, as this now also includes the
"RETURN" keyword, which also needs to be available in an if clause.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>